Renamed some more modules
